Changed the type of client IDs to integers instead of bytes
I don't know if any server will ever get 256 people logged in, but you
never know.

+++ src/net/fe/FEMultiplayer.java
@@ -424,7 +424,7 @@ public class FEMultiplayer extends Game {
 	 *
 	 * @return the players
 	 */
-	public static HashMap<Byte, Player> getPlayers() {
+	public static HashMap<Integer, Player> getPlayers() {
 		return getSession().getPlayerMap();
 	}
