+++ src/net/fe/network/message/KickMessage.java
@@ -9,7 +9,7 @@ public final class KickMessage extends Message {
 	private static final long serialVersionUID = 6558427589809693714L;
 
 	/** The client that is being told to leave */
-	public final byte player;
+	public final int player;
 	/** The reason for the kick */
 	public final String reason;
 
@@ -20,7 +20,7 @@ public final class KickMessage extends Message {
 	 * @param player the leaving client
 	 * @param reason the reason
 	 */
-	public KickMessage(byte origin, byte player, String reason) {
+	public KickMessage(int origin, int player, String reason) {
 		super(origin);
 		this.player = player;
 		this.reason = reason;
